requirements.txt
google-generativeai>=0.7.0
pdfplumber>=0.10.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
rich>=13.0.0